requirements.txt
--find-links=https://girder.github.io/large_image_wheels GDAL
geemap
geopandas
jupyter-server-proxy
leafmap
nbserverproxy
owslib
streamlit
altair<5
streamlit-pandas

